+++ src/gateway_api/models/logic/mod.rs
@@ -1,0 +1,5 @@
+mod request;
+mod response;
+
+pub use request::*;
+pub use response::*;
