+++ src/gateway_api/models/logic/request/mod.rs
@@ -1,0 +1,7 @@
+mod gw_public_key;
+mod state;
+mod transaction;
+
+pub use gw_public_key::*;
+pub use state::*;
+pub use transaction::*;
