+++ src/gateway_api/models/logic/request/state/entity/mod.rs
@@ -1,0 +1,3 @@
+mod state_entity_details;
+
+pub use state_entity_details::*;
