+++ src/gateway_api/models/logic/request/state/entity/state_entity_details.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl StateEntityDetailsRequest {}
