+++ src/gateway_api/models/logic/request/transaction/preview/request_flags.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl TransactionPreviewRequestFlags {}
