+++ src/gateway_api/models/logic/request/transaction/submit/mod.rs
@@ -1,0 +1,3 @@
+mod transaction_submit;
+
+pub use transaction_submit::*;
