+++ src/gateway_api/models/logic/response/ledger_state.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl LedgerState {}
