+++ src/gateway_api/models/logic/response/state/entity/details/fungible/collection.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl FungibleResourcesCollection {}
