+++ src/gateway_api/models/logic/response/state/entity/details/fungible/collection_item.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl FungibleResourcesCollectionItem {}
