+++ src/gateway_api/models/logic/response/state/entity/details/fungible/collection_item_global.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl FungibleResourcesCollectionItemGloballyAggregated {}
