+++ src/gateway_api/models/logic/response/state/entity/details/fungible/mod.rs
@@ -1,0 +1,7 @@
+mod collection;
+mod collection_item;
+mod collection_item_global;
+
+pub use collection::*;
+pub use collection_item::*;
+pub use collection_item_global::*;
