+++ src/gateway_api/models/logic/response/state/entity/details/state_entity_details_response_item.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl StateEntityDetailsResponseItem {}
