+++ src/gateway_api/models/logic/response/state/entity/mod.rs
@@ -1,0 +1,2 @@
+mod details;
+pub use details::*;
