+++ src/gateway_api/models/logic/response/transaction/construction/transaction_construction_response.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl TransactionConstructionResponse {}
