+++ src/gateway_api/models/logic/response/transaction/preview/logs_inner.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl TransactionPreviewResponseLogsInner {}
