+++ src/gateway_api/models/logic/response/transaction/preview/mod.rs
@@ -1,0 +1,5 @@
+mod logs_inner;
+mod transaction_preview_response;
+
+pub use logs_inner::*;
+pub use transaction_preview_response::*;
