+++ src/gateway_api/models/logic/response/transaction/preview/transaction_preview_response.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl TransactionPreviewResponse {}
