+++ src/gateway_api/models/logic/response/transaction/submit/submit.rs
@@ -1,0 +1,3 @@
+use crate::prelude::*;
+
+impl TransactionSubmitResponse {}
