+++ src/gateway_api/models/mod.rs
@@ -1,2 +1,5 @@
+mod logic;
 mod types;
+
+pub use logic::*;
 pub use types::*;
