+++ src/gateway_api/models/types/response/transaction/submit/submit.rs
@@ -11,5 +11,5 @@ use crate::prelude::*;
 )]
 pub struct TransactionSubmitResponse {
     /** Is true if the transaction is a duplicate of an existing pending transaction. */
-    duplicate: bool,
+    pub duplicate: bool,
 }
