+++ src/wrapped_radix_engine_toolkit/low_level/notarized_transaction.rs
@@ -20,6 +20,10 @@ impl NotarizedTransaction {
         })
     }
 
+    pub fn signed_intent(&self) -> &SignedIntent {
+        &self.signed_intent
+    }
+
     pub fn compile(&self) -> CompiledNotarizedIntent {
         compile_notarized_intent(self.clone().into())
         .expect("Should always be able to compile a NotarizedTransaction since we should have called 'compile' in its failing constructor.")
